[RNE Rewrite] refactor(cv): share JSI option-readers, move warpQuad to image_ops
ocr_ops.cpp held two things that aren't OCR-specific:

- The JSI option-readers (getNumberProp/getStringProp/getBoolProp/getBoolPropOr)
  are generic plumbing — promote them to utils.h so image_ops/ocr_ops share one
  copy instead of re-rolling the same hasProperty/isX pattern.
- warpQuad is a generic perspective-crop image op (getPerspectiveTransform +
  warpPerspective + pad/align), no OCR in the math. Move it to image_ops.cpp next
  to resize/cvtColor/gridSample; update headers + install.cpp wiring.

ocr_ops.cpp 835 -> 679 lines; it now holds only OCR detector/sequence
postprocessing (CRAFT grouping, DBNet contouring, CTC argmax). Verified with
check-cpp-warnings.sh (clang++ -fsyntax-only vs ExecuTorch/JSI/OpenCV): clean.
