+++ packages/react-native-executorch/cpp/extensions/cv/image_ops.h
@@ -10,4 +10,8 @@ void install_toChannelsLast(facebook::jsi::Runtime &rt, facebook::jsi::Object &m
 void install_normalize(facebook::jsi::Runtime &rt, facebook::jsi::Object &module);
 void install_applyColormap(facebook::jsi::Runtime &rt, facebook::jsi::Object &module);
 void install_gridSample(facebook::jsi::Runtime &rt, facebook::jsi::Object &module);
+// Perspective-crop an oriented quad of `src` into the `dst` canvas (crop +
+// resize-to-height + pad/align), HWC uint8. Generic image op; the OCR recognizer
+// uses it to lift each text quad into a fixed-height recognizer canvas.
+void install_warpQuad(facebook::jsi::Runtime &rt, facebook::jsi::Object &module);
 } // namespace rnexecutorch::extensions::cv::image_ops
