+++ packages/react-native-executorch/cpp/extensions/cv/install.cpp
@@ -16,12 +16,12 @@ void install(facebook::jsi::Runtime &rt, facebook::jsi::Object &module) {
     image_ops::install_normalize(rt, cvModule);
     image_ops::install_applyColormap(rt, cvModule);
     image_ops::install_gridSample(rt, cvModule);
+    image_ops::install_warpQuad(rt, cvModule);
 
     box_ops::install_nms(rt, cvModule);
     box_ops::install_restrictToBox(rt, cvModule);
 
     ocr_ops::install_extractTextBoxes(rt, cvModule);
-    ocr_ops::install_warpQuad(rt, cvModule);
     ocr_ops::install_ctcGreedyDecode(rt, cvModule);
 
     module.setProperty(rt, "cv", cvModule);
